main script parallel execution

+++ utils/gui_maker.py
@@ -1,7 +1,9 @@
-# TODO: всю отрисовку интерфейса в этот модуль
+# TODO: 1) Необходимо добавить обработку кейса когда нет зареганных пользователей, а так же добавить регистрацию
+#       2)
 
 import pathlib
 from PyQt5 import QtWidgets
+from PyQt5.QtCore import QThread
 from PyQt5.QtGui import QIcon
 from main_standalone_window import Ui_MainWindow
 
@@ -12,9 +14,18 @@ current_file_dir_path = str(pathlib.Path(__file__).parent)
 current_file_dir_parent_path = str(pathlib.Path(__file__).parent.parent)
 
 
+class RunnerWorker(QThread):
+    def __init__(self):
+        QThread.__init__(self)
+
+    def run(self):
+        core_api.main_script_start()
+
+
 class Interface(QtWidgets.QMainWindow):
     def __init__(self):
         super(Interface, self).__init__()
+        self.main_script_runner = RunnerWorker()
         self.ui = Ui_MainWindow()
         self.ui.setupUi(self)
         self.init_ui()
@@ -31,12 +42,13 @@ class Interface(QtWidgets.QMainWindow):
         self.ui.add_link_entry_button.clicked.connect(self.get_link_to_add)
         self.ui.delete_link_entry_button.clicked.connect(self.get_link_to_delete)
         self.ui.time_entry_button.clicked.connect(self.get_time)
-        self.ui.start_script_button.clicked.connect(core_api.main_script_start)
+        self.ui.start_script_button.clicked.connect(self.main_script_runner.start)
 
     def get_link_to_add(self):
         entry_link_value = self.ui.add_link_entry.text()
-        self.ui.add_link_entry.clear()
-        core_api.add_photo(entry_link_value)
+        if entry_link_value:
+            self.ui.add_link_entry.clear()
+            core_api.add_photo(entry_link_value)
 
         # todo: debug
         print(core_api.get_added_photos())
@@ -44,8 +56,9 @@ class Interface(QtWidgets.QMainWindow):
     def get_link_to_delete(self):
         # todo: выводить окно с сообщением удалено или нет
         entry_link_value = self.ui.delete_link_entry.text()
-        self.ui.delete_link_entry.clear()
-        core_api.delete_photo(entry_link_value)
+        if entry_link_value:
+            self.ui.delete_link_entry.clear()
+            core_api.delete_photo(entry_link_value)
 
         # todo: debug
         print(core_api.get_added_photos())
